[validation] some thoughts
